still working on it

+++ app.py
@@ -17,9 +17,9 @@ def index():
 def loan():
     try:
         if request.method == "POST":
-            annual_interest_rate = float(request.form.get("rate"))
-            principal = float(request.form.get("amount"))
-            terms = float(request.form.get("terms"))
+            annual_interest_rate = int(request.form.get("rate"))
+            principal = int(request.form.get("amount"))
+            terms = int(request.form.get("terms"))
 
             # Calculate monthly interest rate
             monthly_interest_rate = annual_interest_rate / 12 / 100
@@ -65,7 +65,7 @@ def savings():
         if request.method == "POST":
             rate = float(request.form.get("rate"))
             starting_amount = float(request.form.get("starting_amount"))
-            terms = float(request.form.get("terms"))
+            terms = int(request.form.get("terms"))
 
             #Determine the number of deposit
             num_deposit = terms * 12
